# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for browsing events and managing bookings.\n\n**Authentication:** call `POST /auth/signin` (or\n`/auth/signup` first, if you don't have an account) to\nget a token. Then click \"Authorize\" (top right) and\nenter it as `Bearer \u003cyour-token\u003e`. Endpoints marked\nwith a padlock require this.\n\n**Errors:** every response uses the same envelope —\n`{ success, message, data, meta, error }`. On failure,\n`error` holds a human-readable message and `data`/`meta`\nare omitted.",
        "title": "AdaMatsuri API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all bookings belonging to the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Get user's booking list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.BookingResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a booking for an event for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a booking",
                "parameters": [
                    {
                        "description": "Booking payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateBookingInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BookingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a booking owned by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.BookingResponse": {
            "type": "object",
            "properties": {
                "booking_code": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/dto.EventResponse"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "dto.EventResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "created_at": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TagResponse"
                    }
                },
                "user": {
                    "$ref": "#/definitions/dto.UserResponse"
                }
            }
        },
...
        "dto.TagResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TagResponse"
                    }
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.Role"
                }
            }
        },
...
        "model.Category": {
            "type": "string",
            "enum": [
                "convention",
                "doujin_market",
                "screening",
                "cosplay_contest",
                "game_tournament",
                "meetup"
            ],
            "x-enum-comments": {
                "CategoryConvention": "large multi-vendor/multi-panel event",
                "CategoryCosplayContest": "cosplay competition or gathering",
                "CategoryDoujinMarket": "artist alley / fan-work sales",
                "CategoryGameTournament": "anime/visual-novel game tournament",
                "CategoryMeetup": "casual fan meetup",
                "CategoryScreening": "anime/movie screening"
            },
            "x-enum-descriptions": [
                "large multi-vendor/multi-panel event",
                "artist alley / fan-work sales",
                "anime/movie screening",
                "cosplay competition or gathering",
                "anime/visual-novel game tournament",
                "casual fan meetup"
            ],
            "x-enum-varnames": [
                "CategoryConvention",
                "CategoryDoujinMarket",
                "CategoryScreening",
                "CategoryCosplayContest",
                "CategoryGameTournament",
                "CategoryMeetup"
            ]
        },
        "model.Role": {
            "type": "string",
            "enum": [
                "attendee",
                "organizer",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleAttendee",
                "RoleOrganizer",
                "RoleAdmin"
            ]
        },
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "service.CreateBookingInput": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer",
                    "example": 1
                },
                "phone": {
                    "type": "string",
                    "example": "+6281234567890"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}